changdu/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "changdu"
version = "0.1.0"
description = "火山方舟 Seedream/Seedance 图像视频生成 CLI"
readme = "README.md"
license = "MIT"
requires-python = ">=3.11"
authors = [
  { name = "bolecodex" },
]
keywords = ["seedream", "seedance", "volcano-ark", "image-generation", "video-generation", "cli"]
classifiers = [
  "Development Status :: 4 - Beta",
  "Environment :: Console",
  "Intended Audience :: Developers",
  "Programming Language :: Python :: 3.11",
  "Programming Language :: Python :: 3.12",
  "Programming Language :: Python :: 3.13",
  "Topic :: Multimedia :: Graphics",
  "Topic :: Multimedia :: Video",
]
dependencies = [
  "httpx>=0.27.0",
  "pydantic>=2.7.0",
  "pyyaml>=6.0.1",
  "typer>=0.12.3",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.2.0",
]

[project.scripts]
changdu = "changdu.cli.main:main"

[project.urls]
Homepage = "https://github.com/bolecodex/changdu-skills"
Repository = "https://github.com/bolecodex/changdu-skills"
Issues = "https://github.com/bolecodex/changdu-skills/issues"

[tool.pytest.ini_options]
testpaths = ["tests"]
